Move model to specific package.
