+++ app/src/main/java/com/preguardia/app/consultation/approve/ApproveConsultationPresenter.java
@@ -10,7 +10,7 @@ import com.google.common.base.Joiner;
 import com.orhanobut.logger.Logger;
 import com.preguardia.app.BuildConfig;
 import com.preguardia.app.R;
-import com.preguardia.app.consultation.model.Consultation;
+import com.preguardia.app.data.model.Consultation;
 import com.preguardia.app.general.Constants;
 import com.preguardia.app.user.model.Medic;
 import com.preguardia.app.user.model.Patient;
