+++ app/src/main/java/com/preguardia/app/consultation/details/ConsultationDetailsPresenter.java
@@ -9,8 +9,8 @@ import com.firebase.client.FirebaseError;
 import com.firebase.client.ValueEventListener;
 import com.orhanobut.logger.Logger;
 import com.preguardia.app.BuildConfig;
-import com.preguardia.app.consultation.model.Consultation;
-import com.preguardia.app.consultation.model.GenericMessage;
+import com.preguardia.app.data.model.Consultation;
+import com.preguardia.app.data.model.GenericMessage;
 import com.preguardia.app.general.Constants;
 import com.preguardia.app.user.model.Medic;
 import com.preguardia.app.user.model.Patient;
