+++ app/src/main/java/com/preguardia/app/data/model/Question.java
@@ -1,4 +1,4 @@
-package com.preguardia.app.consultation.model;
+package com.preguardia.app.data.model;
 
 /**
  * @author amouly on 3/1/16.
